refactor: clean up imports and folder structure
