+++ packages/automated-dispute/src/external.ts
@@ -1,0 +1,1 @@
+export { EboProcessor, EboActorsManager } from "./services/index.js";
