+++ packages/automated-dispute/src/index.ts
@@ -1,1 +1,1 @@
-export * from "./protocolProvider.js";
+export * from "./external.js";
