+++ packages/automated-dispute/src/providers/protocolProvider.ts
@@ -19,17 +19,17 @@ import {
 import { privateKeyToAccount } from "viem/accounts";
 import { arbitrum } from "viem/chains";
 
-import type { EboEvent, EboEventName } from "./types/events.js";
-import type { Dispute, Request, Response } from "./types/prophet.js";
-import { eboRequestCreatorAbi, epochManagerAbi, oracleAbi } from "./abis/index.js";
-import { RpcUrlsEmpty } from "./exceptions/rpcUrlsEmpty.exception.js";
+import type { EboEvent, EboEventName } from "../types/events.js";
+import type { Dispute, Request, Response } from "../types/prophet.js";
+import { eboRequestCreatorAbi, epochManagerAbi, oracleAbi } from "../abis/index.js";
+import { RpcUrlsEmpty } from "../exceptions/rpcUrlsEmpty.exception.js";
 import {
     IProtocolProvider,
     IReadProvider,
     IWriteProvider,
     ProtocolContractsAddresses,
-} from "./interfaces/index.js";
-import { ErrorFactory } from "./services/errorFactory.js";
+} from "../interfaces/index.js";
+import { ErrorFactory } from "../services/errorFactory.js";
 
 // TODO: these constants should be env vars
 const TRANSACTION_RECEIPT_CONFIRMATIONS = 1;
