+++ packages/automated-dispute/src/services/index.ts
@@ -1,3 +1,5 @@
-export * from "../eboActor.js";
+export * from "./eboActor.js";
+export * from "./eboActorsManager.js";
 export * from "./eboProcessor.js";
 export * from "./eboRegistry/index.js";
+export * from "./errorFactory.js";
