+++ packages/automated-dispute/tests/services/eboActor/fixtures.ts
@@ -1,6 +1,6 @@
 import { Address } from "viem";
 
-import { Request, RequestId } from "../../src/types/prophet";
+import { Request, RequestId } from "../../../src/types/index.js";
 
 export const mockedPrivateKey =
     "0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80";
