+++ packages/automated-dispute/tests/services/eboActor/onDisputeStatusChanged.spec.ts
@@ -1,8 +1,8 @@
 import { ILogger } from "@ebo-agent/shared";
 import { describe, expect, it, vi } from "vitest";
 
-import { EboEvent } from "../../src/types/events";
-import mocks from "../mocks/index.ts";
+import { EboEvent } from "../../../src/types/index.js";
+import mocks from "../../mocks/index.ts";
 import { DEFAULT_MOCKED_REQUEST_CREATED_DATA } from "./fixtures.ts";
 
 const logger: ILogger = {
