+++ packages/automated-dispute/tests/services/errorFactory.spec.ts
@@ -1,10 +1,12 @@
 import { describe, expect, it } from "vitest";
 
-import { EBORequestCreator_ChainNotAdded } from "../src/exceptions/chainNotAdded.exception.js";
-import { EBORequestCreator_InvalidEpoch } from "../src/exceptions/invalidEpoch.exception.js";
-import { Oracle_InvalidRequestBody } from "../src/exceptions/invalidRequestBody.exception.js";
-import { EBORequestModule_InvalidRequester } from "../src/exceptions/invalidRequester.exception.js";
-import { ErrorFactory } from "../src/services/errorFactory.js";
+import {
+    EBORequestCreator_ChainNotAdded,
+    EBORequestCreator_InvalidEpoch,
+    EBORequestModule_InvalidRequester,
+    Oracle_InvalidRequestBody,
+} from "../../src/exceptions/index.js";
+import { ErrorFactory } from "../../src/services/index.js";
 
 describe("ErrorFactory", () => {
     it("creates EBORequestCreator_InvalidEpoch error", () => {
